refactor(types): single-home AgentUsage in the SDK, drop the plugin-dir-escaping import
AgentUsage was declared in src/core/agent-usage.ts, copied verbatim into
the health plugin's health-page.tsx, and imported by the team plugin's
overview-tab via a relative path that escaped the plugin directory into
src/core (the only client file in any plugin reaching outside its own
tree). It's now a type-only export from the SDK (the usage-panel wire
shape); src/core/agent-usage.ts re-exports it for server callers, and
both client components import it from @makinbakin/sdk/types. The verbatim
copy and the escaping import are gone. Audit finding: sdk-gaps.
